requirements.txt
Werkzeug==2.0.2
Flask_SQLAlchemy==2.5.1
flask_marshmallow==0.14.0
Flask_Migrate==3.1.0
Flask==2.0.2
requests==2.26.0
Flask_Testing==0.8.1
SQLAlchemy==1.4.25
aiohttp==3.7.4.post0
PyJWT==2.3.0
